Separation similar function in Scroller.
