+++ Assets/Scenes/SlotTurnOver.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 586241f05d0474a4eb500573c274bd11
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
