+++ Assets/Scripts/Game/UI/Scroller/Core.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 5eae3ae02a2048168eebb4bb87a8a272
+timeCreated: 1596412826
