+++ Assets/Scripts/Game/UI/Scroller/Core/ScrollContent.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using UnityEngine;
+
+namespace Game.UI.Scroller.Core
+{
+    public static class ScrollContent
+    {
+        [Serializable]
+        public class ScrollContentData
+        {
+            public ScrollContentData()
+            {
+            }
+        }
+    }
+
+    public class ScrollContent<TData> : MonoBehaviour where TData : ScrollContent.ScrollContentData
+    {
+        private TData _data;
+
+        public virtual void SetContentData(TData tdata)
+        {
+            _data = tdata;
+        }
+    }
+}
