+++ Assets/Scripts/Game/UI/Scroller/Core/Scroller.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 57939bd4f9b54d428a89109d277f492f
+timeCreated: 1596412816
